Changement sur l'interaction avec la BD changement dans la structure
