+++ .gitignore
@@ -1,0 +1,2 @@
+cache/*
+templates_c/*
