+++ application/libs/Utils.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/**
+ * Description of utils
+ *
+ * @author Moussa
+ */
+class Utils {
+   
+    private static function randomString($length = 50)
+    {
+	$characters = '0123456789abcdefghijklmnopqrstuvwxyz';
+	$string = '';    
+		
+	for ($p = 0; $p < $length; $p++) {
+		$string .= $characters[mt_rand(0, strlen($characters))];
+	}
+		
+      	return $string;
+    }
+
+
+}
