+++ system/config.php
@@ -24,4 +24,4 @@ define('WEBAPPROOT', WEBROOT . 'application/'); //application dossier
 define('DB_HOST', 'localhost');
 define('DB_USER', 'root');
 define('DB_PASSWORD', 'root');
-define('DB_DB', 'rpm_bd');
+define('DB_DB', 'bd_rpm');
